+++ src/output/wrappers.rs
@@ -1,0 +1,4 @@
+//! Some types do not appear "nicely" when serialized
+//! This module provides wrappers for them that implements custom serialization and deserialization
+
+pub mod bool;
